♻️ Separate GitHub and GitLab logic
